# apps/schulapp-backend/openapi.yaml
openapi: "3.0.3"
info:
  title: SchulApp API
  description: API für die Schulverwaltungs-App
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Lokale Entwicklung
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    User:
      type: object
      required: [id, email, first_name, last_name, role]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [student, teacher, school_admin, superadmin]
        school_id:
          type: integer
          nullable: true
...
    Class:
      type: object
      required: [id, name, school_year]
      properties:
        id:
          type: integer
        name:
          type: string
        school_year:
          type: string
        school_id:
          type: integer
        created_at:
          type: string
          format: date-time
...
    RegisterRequest:
      type: object
      required: [email, password, first_name, last_name, school_id]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        first_name:
          type: string
        last_name:
          type: string
        school_id:
          type: integer
        requested_class_id:
          type: integer
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Öffentliche Schüler-Registrierung (legt Konto an + optional Beitrittsanfrage)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registriert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: E-Mail bereits vergeben
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/public/classes:
    get:
      tags: [public]
      summary: Klassen einer Schule öffentlich auflisten (für Registrierung)
      parameters:
        - name: school_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Klassenliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
  /api/v1/schools:
    get:
      tags: [schools]
      summary: Schulen auflisten
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Schulliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      tags: [schools]
      summary: Schule anlegen (nur superadmin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolRequest"
      responses:
        "201":
          description: Schule angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /api/v1/me/membership:
    get:
      tags: [auth]
      summary: Eigene Schul-Zugehörigkeit (Schule, Klassen, offene Anfragen)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Zugehörigkeit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
...
  /api/v1/users:
    get:
      tags: [users]
      summary: Alle Benutzer auflisten
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Benutzerliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags: [users]
      summary: Benutzer anlegen
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Benutzer angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/v1/users/bulk:
    post:
      tags: [users]
      summary: Mehrere Benutzer anlegen
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateUsersRequest"
      responses:
        "201":
          description: Benutzer angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCreateUsersResponse"
  /api/v1/users/{id}:
    get:
      tags: [users]
      summary: Benutzer abrufen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Benutzer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      tags: [users]
      summary: Benutzer aktualisieren
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Aktualisierter Benutzer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags: [users]
      summary: Benutzer löschen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Benutzer gelöscht
  /api/v1/classes:
    get:
      tags: [classes]
      summary: Klassen auflisten
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Klassenliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      tags: [classes]
      summary: Klasse anlegen
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Klasse angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /api/v1/classes/{id}:
    get:
      tags: [classes]
      summary: Klasse abrufen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Klasse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    patch:
      tags: [classes]
      summary: Klasse aktualisieren
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassRequest"
      responses:
        "200":
          description: Aktualisierte Klasse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      tags: [classes]
      summary: Klasse löschen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Klasse gelöscht
  /api/v1/classes/{id}/members:
    get:
      tags: [classes]
      summary: Klassenmitglieder auflisten
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mitgliederliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassMember"
    post:
      tags: [classes]
      summary: Mitglied hinzufügen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Mitglied hinzugefügt
  /api/v1/classes/{id}/members/{userId}:
    delete:
      tags: [classes]
      summary: Mitglied entfernen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Mitglied entfernt
  /api/v1/classes/{id}/join-requests:
    get:
      tags: [classes]
      summary: Beitrittsanfragen einer Klasse auflisten
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Anfragenliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JoinRequest"
    post:
      tags: [classes]
      summary: Beitrittsanfrage stellen (Schüler)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Anfrage gestellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequest"
  /api/v1/join-requests/{id}/approve:
    post:
      tags: [classes]
      summary: Beitrittsanfrage freigeben
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Freigegeben
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequest"
  /api/v1/join-requests/{id}/reject:
    post:
      tags: [classes]
      summary: Beitrittsanfrage ablehnen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Abgelehnt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequest"
  /api/v1/classes/{id}/teachers:
    get:
      tags: [classes]
      summary: Klassenlehrer auflisten
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lehrerliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassTeacher"
    post:
      tags: [classes]
      summary: Lehrer zuweisen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTeacherRequest"
      responses:
        "201":
          description: Lehrer zugewiesen
  /api/v1/classes/{id}/teachers/{userId}:
    delete:
      tags: [classes]
      summary: Lehrer entfernen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Lehrer entfernt
  /api/v1/classes/{id}/homework:
    get:
      tags: [homework]
      summary: Hausaufgaben einer Klasse auflisten
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hausaufgabenliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Homework"
    post:
      tags: [homework]
      summary: Hausaufgabe anlegen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHomeworkRequest"
      responses:
        "201":
          description: Hausaufgabe angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Homework"
  /api/v1/classes/{id}/lessons:
    get:
      tags: [lessons]
      summary: Stundenplan einer Klasse auflisten (recurring Lessons, optional mit Vertretungen einer Woche überlagert)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: week_of
          in: query
          description: "Beliebiges Datum innerhalb der gewünschten Woche (YYYY-MM-DD). Lessons werden für diese Kalenderwoche auf Daten projiziert und mit Substitutions derselben Klasse+Datum+Stunde zusammengeführt. Ohne Angabe wird die aktuelle Woche verwendet."
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Stundenplan der Woche, Lessons überlagert mit Substitutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableEntry"
    post:
      tags: [lessons]
      summary: Stundenplan-Slot anlegen (teacher der eigenen Klasse, school_admin/superadmin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonRequest"
      responses:
        "201":
          description: Stundenplan-Slot angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /api/v1/lessons/{id}:
    patch:
      tags: [lessons]
      summary: Stundenplan-Slot aktualisieren
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonRequest"
      responses:
        "200":
          description: Aktualisierter Stundenplan-Slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      tags: [lessons]
      summary: Stundenplan-Slot löschen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Stundenplan-Slot gelöscht
  /api/v1/homework/{id}:
    patch:
      tags: [homework]
      summary: Hausaufgabe aktualisieren
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHomeworkRequest"
      responses:
        "200":
          description: Aktualisierte Hausaufgabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Homework"
    delete:
      tags: [homework]
      summary: Hausaufgabe löschen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Hausaufgabe gelöscht
  /api/v1/homework/{id}/submissions:
    get:
      tags: [homework]
      summary: Abgaben einer Hausaufgabe auflisten
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Abgabenliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HomeworkSubmission"
    post:
      tags: [homework]
      summary: Hausaufgabe abgeben
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Abgabe erfolgreich
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkSubmission"
  /api/v1/submissions/{id}:
    patch:
      tags: [homework]
      summary: Abgabe bewerten
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeSubmissionRequest"
      responses:
        "200":
          description: Bewertung gespeichert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkSubmission"
    delete:
      tags: [homework]
      summary: Abgabe zurückziehen (nur solange nicht benotet)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Zurückgezogen
  /api/v1/substitutions:
    get:
      tags: [substitutions]
      summary: Vertretungen auflisten
      security:
        - bearerAuth: []
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: class_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Vertretungsliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Substitution"
    post:
      tags: [substitutions]
      summary: Vertretung anlegen
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubstitutionRequest"
      responses:
        "201":
          description: Vertretung angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Substitution"
...
  /api/v1/folders/{id}:
    delete:
      tags: [files]
      summary: Ordner löschen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Ordner gelöscht
...
  /api/v1/chat/contacts:
    get:
      tags: [chat]
      summary: Erreichbare Personen für einen Chat auflisten
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Kontakte aus gemeinsamen Klassen; Admins sehen alle aktiven Konten
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatContact"
...